[REFACTOR] ids type from Integer to Long

+++ src/main/java/com/pvt/SocialSips/questpool/QuestpoolRepository.java
@@ -2,6 +2,6 @@ package com.pvt.SocialSips.questpool;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestpoolRepository extends JpaRepository<Questpool, Integer> {
+public interface QuestpoolRepository extends JpaRepository<Questpool, Long> {
 
 }
